requirements.txt
aniso8601==9.0.1
attrs==23.1.0
backoff==2.2.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
Deprecated==1.2.14
Flask==1.0.2
flask-restplus==0.12.1
googleapis-common-protos==1.56.2
grpcio==1.59.3
gunicorn==20.0.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==1.1.0
Jinja2==3.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
MarkupSafe==2.1.3
memproxy==0.3.0rc19
msgspec==0.18.4
mypy==1.7.0
mypy-extensions==1.0.0
opentelemetry-api==1.14.0
opentelemetry-exporter-jaeger==1.14.0
opentelemetry-exporter-jaeger-proto-grpc==1.14.0
opentelemetry-exporter-jaeger-thrift==1.14.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-instrumentation==0.35b0
opentelemetry-instrumentation-flask==0.35b0
opentelemetry-instrumentation-wsgi==0.35b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.14.0
opentelemetry-semantic-conventions==0.35b0
opentelemetry-util-http==0.35b0
packaging==23.2
pkgutil-resolve-name==1.3.10
prometheus-client==0.18.0
prometheus-flask-exporter==0.11.0
protobuf==3.20.3
pycparser==2.21
pytz==2023.3.post1
redis==3.2.0
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.0
six==1.16.0
thrift==0.16.0
tomli==2.0.1
types-pyOpenSSL==23.3.0.0
types-redis==4.6.0.10
typing-extensions==4.8.0
urllib3==2.1.0
Werkzeug==0.15.6
wrapt==1.16.0
zipp==3.17.0
